Move the OpenTelemetry formatter to the extension rather than core Envoy

t #
