+++ source/common/formatter/substitution_formatter.h
@@ -14,7 +14,6 @@
 
 #include "absl/container/flat_hash_map.h"
 #include "absl/types/optional.h"
-#include "opentelemetry/proto/common/v1/common.pb.h"
 
 namespace Envoy {
 namespace Formatter {
@@ -241,75 +240,6 @@ private:
 };
 
 using StructFormatterPtr = std::unique_ptr<StructFormatter>;
-
-/**
- * A formatter for OpenTelemetry logs, which returns a KeyValueList proto.
- */
-class OpenTelemetryFormatter {
-public:
-  OpenTelemetryFormatter(const ::opentelemetry::proto::common::v1::KeyValueList& format_mapping);
-
-  ::opentelemetry::proto::common::v1::KeyValueList
-  format(const Http::RequestHeaderMap& request_headers,
-         const Http::ResponseHeaderMap& response_headers,
-         const Http::ResponseTrailerMap& response_trailers,
-         const StreamInfo::StreamInfo& stream_info, absl::string_view local_reply_body) const;
-
-private:
-  struct OpenTelemetryFormatMapWrapper;
-  struct OpenTelemetryFormatListWrapper;
-  using OpenTelemetryFormatValue =
-      absl::variant<const std::vector<FormatterProviderPtr>, const OpenTelemetryFormatMapWrapper,
-                    const OpenTelemetryFormatListWrapper>;
-  // We use std::list to keep the order of the repeated "values" field in the OpenTelemetry
-  // KeyValueList.
-  using OpenTelemetryFormatMap = std::list<std::pair<std::string, OpenTelemetryFormatValue>>;
-  using OpenTelemetryFormatMapPtr = std::unique_ptr<OpenTelemetryFormatMap>;
-  struct OpenTelemetryFormatMapWrapper {
-    OpenTelemetryFormatMapPtr value_;
-  };
-
-  using OpenTelemetryFormatList = std::list<OpenTelemetryFormatValue>;
-  using OpenTelemetryFormatListPtr = std::unique_ptr<OpenTelemetryFormatList>;
-  struct OpenTelemetryFormatListWrapper {
-    OpenTelemetryFormatListPtr value_;
-  };
-
-  using OpenTelemetryFormatMapVisitor = StructFormatMapVisitorHelper<
-      const std::function<::opentelemetry::proto::common::v1::AnyValue(
-          const std::vector<FormatterProviderPtr>&)>,
-      const std::function<::opentelemetry::proto::common::v1::AnyValue(
-          const OpenTelemetryFormatter::OpenTelemetryFormatMapWrapper&)>,
-      const std::function<::opentelemetry::proto::common::v1::AnyValue(
-          const OpenTelemetryFormatter::OpenTelemetryFormatListWrapper&)>>;
-
-  // Methods for building the format map.
-  class FormatBuilder {
-  public:
-    std::vector<FormatterProviderPtr> toFormatStringValue(const std::string& string_format) const;
-    OpenTelemetryFormatMapWrapper
-    toFormatMapValue(const ::opentelemetry::proto::common::v1::KeyValueList& struct_format) const;
-    OpenTelemetryFormatListWrapper toFormatListValue(
-        const ::opentelemetry::proto::common::v1::ArrayValue& list_value_format) const;
-  };
-
-  // Methods for doing the actual formatting.
-  ::opentelemetry::proto::common::v1::AnyValue
-  providersCallback(const std::vector<FormatterProviderPtr>& providers,
-                    const Http::RequestHeaderMap& request_headers,
-                    const Http::ResponseHeaderMap& response_headers,
-                    const Http::ResponseTrailerMap& response_trailers,
-                    const StreamInfo::StreamInfo& stream_info,
-                    absl::string_view local_reply_body) const;
-  ::opentelemetry::proto::common::v1::AnyValue openTelemetryFormatMapCallback(
-      const OpenTelemetryFormatter::OpenTelemetryFormatMapWrapper& format_map,
-      const OpenTelemetryFormatMapVisitor& visitor) const;
-  ::opentelemetry::proto::common::v1::AnyValue openTelemetryFormatListCallback(
-      const OpenTelemetryFormatter::OpenTelemetryFormatListWrapper& format_list,
-      const OpenTelemetryFormatMapVisitor& visitor) const;
-
-  const OpenTelemetryFormatMapWrapper kv_list_output_format_;
-};
 
 class JsonFormatterImpl : public Formatter {
 public:
