+++ source/extensions/access_loggers/open_telemetry/BUILD
@@ -34,9 +34,9 @@ envoy_cc_library(
     hdrs = ["access_log_impl.h"],
     deps = [
         ":grpc_access_log_lib",
+        ":substitution_formatter_lib",
         "//envoy/access_log:access_log_interface",
         "//envoy/protobuf:message_validator_interface",
-        "//source/common/formatter:substitution_formatter_lib",
         "//source/common/protobuf:utility_lib",
         "//source/extensions/access_loggers/common:access_log_base",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
@@ -70,5 +70,18 @@ envoy_cc_extension(
         "//source/extensions/access_loggers/open_telemetry:access_log_proto_descriptors_lib",
         "//source/extensions/access_loggers/open_telemetry:grpc_access_log_lib",
         "@envoy_api//envoy/extensions/access_loggers/open_telemetry/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "substitution_formatter_lib",
+    srcs = ["substitution_formatter.cc"],
+    hdrs = ["substitution_formatter.h"],
+    external_deps = ["abseil_str_format"],
+    deps = [
+        "//envoy/stream_info:stream_info_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/formatter:substitution_formatter_lib",
+        "@opentelemetry_proto//:common_cc_proto",
     ],
 )
