+++ source/extensions/access_loggers/open_telemetry/access_log_impl.h
@@ -11,10 +11,10 @@
 #include "envoy/singleton/instance.h"
 #include "envoy/thread_local/thread_local.h"
 
-#include "source/common/formatter/substitution_formatter.h"
 #include "source/common/grpc/typed_async_client.h"
 #include "source/extensions/access_loggers/common/access_log_base.h"
 #include "source/extensions/access_loggers/open_telemetry/grpc_access_log_impl.h"
+#include "source/extensions/access_loggers/open_telemetry/substitution_formatter.h"
 
 #include "opentelemetry/proto/collector/logs/v1/logs_service.pb.h"
 #include "opentelemetry/proto/common/v1/common.pb.h"
@@ -56,8 +56,8 @@ private:
 
   const ThreadLocal::SlotPtr tls_slot_;
   const GrpcAccessLoggerCacheSharedPtr access_logger_cache_;
-  std::unique_ptr<Formatter::OpenTelemetryFormatter> body_formatter_;
-  std::unique_ptr<Formatter::OpenTelemetryFormatter> attributes_formatter_;
+  std::unique_ptr<OpenTelemetryFormatter> body_formatter_;
+  std::unique_ptr<OpenTelemetryFormatter> attributes_formatter_;
 };
 
 using AccessLogPtr = std::unique_ptr<AccessLog>;
