+++ test/extensions/access_loggers/open_telemetry/BUILD
@@ -1,5 +1,7 @@
 load(
     "//bazel:envoy_build_system.bzl",
+    "envoy_benchmark_test",
+    "envoy_cc_benchmark_binary",
     "envoy_package",
 )
 load(
@@ -80,4 +82,42 @@ envoy_extension_cc_test(
         "@envoy_api//envoy/extensions/filters/network/http_connection_manager/v3:pkg_cc_proto",
         "@opentelemetry_proto//:logs_cc_proto",
     ],
+)
+
+envoy_extension_cc_test(
+    name = "substitution_formatter_test",
+    srcs = ["substitution_formatter_test.cc"],
+    extension_names = ["envoy.access_loggers.open_telemetry"],
+    deps = [
+        "//source/common/formatter:substitution_formatter_lib",
+        "//source/common/http:exception_lib",
+        "//source/common/http:header_map_lib",
+        "//source/common/router:string_accessor_lib",
+        "//source/extensions/access_loggers/open_telemetry:substitution_formatter_lib",
+        "//test/mocks/stream_info:stream_info_mocks",
+        "//test/mocks/upstream:cluster_info_mocks",
+        "//test/test_common:utility_lib",
+        "@opentelemetry_proto//:logs_cc_proto",
+    ],
+)
+
+envoy_cc_benchmark_binary(
+    name = "substitution_formatter_speed_test",
+    srcs = ["substitution_formatter_speed_test.cc"],
+    external_deps = [
+        "benchmark",
+    ],
+    deps = [
+        "//source/common/http:header_map_lib",
+        "//source/common/network:address_lib",
+        "//source/extensions/access_loggers/open_telemetry:substitution_formatter_lib",
+        "//test/common/stream_info:test_util",
+        "//test/mocks/stream_info:stream_info_mocks",
+        "@opentelemetry_proto//:common_cc_proto",
+    ],
+)
+
+envoy_benchmark_test(
+    name = "substitution_formatter_speed_test_benchmark_test",
+    benchmark_binary = "substitution_formatter_speed_test",
 )
